linkedin-automation/requirements.txt
# Python dependencies for LinkedIn Automation & Job Search Toolkit
# Install: pip install -r requirements.txt

# Core Dependencies
python-dotenv==1.0.1
requests==2.32.3

# Browser Automation (REQUIRED for company page posting)
playwright==1.47.0

# Image Processing (for carousels and covers)
Pillow==10.4.0

# PDF Processing
pdfplumber==0.11.4
pypdf==5.1.0

# Document Generation
python-docx==1.1.2

# Web Scraping
beautifulsoup4==4.12.3
selenium==4.25.0

# NLP & Text Processing (optional)
spacy==3.8.2
# Download model: python -m spacy download en_core_web_sm

# Data Processing
pandas==2.2.3
numpy==2.1.2

# Database
sqlalchemy==2.0.35

# Encryption
cryptography==43.0.1

# AI Content Generation (optional)
openai==1.51.2

# Progress Bars
tqdm==4.66.5

# Date/Time
python-dateutil==2.9.0

# Markdown
markdown==3.7
